Refactor kafka topic
