+++ src/Kafka/KafkaTopics/kafkaTopics.stories.tsx
@@ -1,10 +1,6 @@
-import { action } from "@storybook/addon-actions";
 import type { ComponentMeta, ComponentStory } from "@storybook/react";
-import { fakeApi } from "../../shared/storiesHelpers";
-import { KafkaTopics as KafkaTopicsComp } from "./KafkaTopics";
 import type { KafkaTopic } from "./types";
-
-const getTopicsAction = action("getTopics");
+import { KafkaTopics } from "./KafkaTopics";
 
 const topics: KafkaTopic[] = [
   {
@@ -28,48 +24,23 @@ const topics: KafkaTopic[] = [
 ];
 
 export default {
-  component: KafkaTopicsComp,
+  component: KafkaTopics,
   args: {
-    getUrlFortopic: (row) => `/${row.topic_name}`,
+    getUrlFortopic: () => "",
+    topics: topics,
+    topicName: [],
   },
-} as ComponentMeta<typeof KafkaTopicsComp>;
+} as ComponentMeta<typeof KafkaTopics>;
 
-const randomData = new Array(50).fill(0).map((_, index) => {
-  const i = topics[index % topics.length];
-  return {
-    ...i,
-    topic_name: `${i.topic_name}-${index}`,
-  };
-});
+const Template: ComponentStory<typeof KafkaTopics> = (args) => (
+  <KafkaTopics {...args} />
+);
 
-const Template: ComponentStory<typeof KafkaTopicsComp> = (args) => {
-  return (
-    <KafkaTopicsComp
-      {...args}
-      getTopics={(page, perPage, query, sort, sortDirection) => {
-        const allData = randomData.filter((i) => {
-          return query.names.length > 0
-            ? query.names.find((n) => i.topic_name.includes(n))
-            : true;
-        });
-        const data = allData.slice(page, perPage);
-        const count = allData.length;
-        getTopicsAction(
-          ...Object.entries({
-            page,
-            perPage,
-            sort,
-            sortDirection,
-          }).flat(),
-          ...Object.entries(query)
-            .map(([k, v]) => [k, v.length > 0 ? v.join(",") : null])
-            .flat()
-        );
-        return fakeApi({ topics: data, count });
-      }}
-    />
-  );
+export const KafkaTopicsTable = Template.bind({});
+KafkaTopicsTable.args = {};
+
+export const NoTopics = Template.bind({});
+NoTopics.args = {
+  topics: [],
+  topicName: ["boo", "foo"],
 };
-
-export const KafkaTopics = Template.bind({});
-KafkaTopics.args = {};
